Move service worker client to its own package
