+++ packages/core/integration-tests/test/integration/service-worker/manifest-worker.js
@@ -1,3 +1,3 @@
-import {manifest, version} from '@parcel/runtime-service-worker';
+import {manifest, version} from '@parcel/service-worker';
 
 output(manifest, version);
